+++ assets/javascripts/kitten/components/structure/cards/project-card/stories.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { ProjectCard } from './index'
 import { DocsPage } from 'storybook/docs-page'
 import { Text, Button, Tag, FlexWrapper, LockOutlineIcon } from 'kitten'
-import Truncate from 'react-truncate'
 
 export default {
   title: 'Structure/Cards/ProjectCard',
@@ -21,7 +20,10 @@ export default {
 
 const argTypes = {
   status: { control: 'select' },
-  progressProps: { control: 'object' },
+  progressProps: {
+    name: 'Story component: progressProps',
+    control: 'object',
+  },
 }
 
 const args = {
@@ -69,10 +71,10 @@ export const Default = ({ progressProps, ...args }) => (
         </Text>
       </ProjectCard.Item>
       <ProjectCard.Item>
-        <Text className="k-u-ellipsis" weight="bold" size="tiny">
+        <Text className="k-u-nowrap" weight="bold" size="tiny">
           9 930 €
         </Text>
-        <Text className="k-u-ellipsis" size="micro">
+        <Text className="k-u-nowrap" size="micro">
           sur 12 000 €
         </Text>
       </ProjectCard.Item>
@@ -239,12 +241,10 @@ export const PermanentProjects = ({ progressProps, ...args }) => (
         The Office, la série culte décortiquée par S!CK
       </ProjectCard.Title>
       <ProjectCard.Line>
-        <Text size="micro">
-          <Truncate lines={2}>
-            Ne laisser personne pour compte : plus qu’une devise. Un objectif et
-            une valeur de Un objectif et une valeur de Un objectif et une valeur
-            de
-          </Truncate>
+        <Text size="micro" className="k-u-clamp-2">
+          Ne laisser personne pour compte : plus qu’une devise. Un objectif et
+          une valeur de Un objectif et une valeur de Un objectif et une valeur
+          de
         </Text>
       </ProjectCard.Line>
       <ProjectCard.Line>